+++ model.py
@@ -1,22 +1,64 @@
-import torch
 import torch.nn as nn
+import torch
+import numpy as np
 
-class Model(nn.Module):
 
-    def __init__(self, input_size):
+class Res_block(nn.Module):
 
-        super(Model, self).__init__()
-        self.MLP = nn.Sequential()
-        self.MLP.add_module('layer1', nn.Linear(input_size, 128))
-        self.MLP.add_module('relu1', nn.ReLU(True))
-        self.MLP.add_module('layer2', nn.Linear(128, 256))
-        self.MLP.add_module('relu2', nn.ReLU(True))
-        self.MLP.add_module('layer3', nn.Linear(256, 256))
-        self.MLP.add_module('relu3', nn.ReLU(True))
-        self.MLP.add_module('output', nn.Linear(256, 1))
+    def __init__(self, inplanes, kernel_size, padding):
+        super(Res_block, self).__init__()
 
-    def forward(self, input):
+        self.conv1 = nn.Conv2d(inplanes, inplanes,
+                               kernel_size, padding=padding)
+        self.batchn1 = nn.BatchNorm2d(inplanes)
+        self.conv2 = nn.Conv2d(inplanes, inplanes,
+                               kernel_size, padding=padding)
+        self.batchn2 = nn.BatchNorm2d(inplanes)
+
+        self.relu = nn.ReLU(inplace=True)
+
+    def forward(self, input_x):
+
+        identity = input_x
+
+        out = self.conv1(input_x)
+        out = self.batchn1(out)
+        out = self.relu(out)
+
+        out = self.conv2(out)
+        out = self.batchn2(out)
+
         
-        return self.MLP(input)
+        out += identity
+        out = self.relu(out)
+
+        return out
+
+class ResModel(nn.Module):
+
+    def __init__(self, input_shape, num_layers, kernel_size, 
+            num_of_res_layers, padding, number_of_filters):
+        super(ResModel, self).__init__()
+
+
+        self.model = nn.Sequential()
+        self.model.add_module('Conv Start', 
+                    nn.Conv2d(num_layers, number_of_filters, kernel_size=kernel_size, padding=padding))
         
+        self.model.add_module('Batch_norm', nn.BatchNorm2d(number_of_filters))
+        self.model.add_module('relu', nn.ReLU(inplace=True))
+
+        for i in range(0,num_of_res_layers):
+            self.model.add_module('Res_Layer {}'.format(i), 
+                    Res_block(number_of_filters, kernel_size, padding))
+
+        self.model.add_module('Conv Flat', nn.Conv2d(number_of_filters, 8, kernel_size=(1, 1)))
+        self.model.add_module('Flatten', nn.Flatten())
+        self.model.add_module('Output', nn.Linear(8 * input_shape[0] * input_shape[1], 1))
+
+
+    def forward(self, input_x):
+
+        return self.model(input_x)
+
